serverless/mcp-sample/native_lambda/demo-mcp/requirements-layer.txt
mcp
strands-agents
strands-agents-tools